+++ src/DynamoCore/Engine/EngineController.cs
@@ -1,15 +1,11 @@
 ﻿using Autodesk.DesignScript.Interfaces;
 using Dynamo.Engine.CodeCompletion;
-using Dynamo.Core.Threading;
-using Dynamo.Interfaces;
 using Dynamo.Models;
-using Dynamo.Nodes;
 using Dynamo.Logging;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using ProtoCore.AST.AssociativeAST;
 using ProtoCore.DSASM.Mirror;
 using ProtoCore.Mirror;
-using ProtoCore.Namespace;
 using ProtoScript.Runners;
 using System;
 using System.Collections.Generic;
