+++ src/DynamoCore/Models/DynamoModel.cs
@@ -11,7 +11,6 @@ using System.Threading;
 using System.Xml;
 using Dynamo.Configuration;
 using Dynamo.Core;
-using Dynamo.Core.Threading;
 using Dynamo.Engine;
 using Dynamo.Extensions;
 using Dynamo.Interfaces;
@@ -19,8 +18,7 @@ using Dynamo.Migration;
 using Dynamo.Models.NodeLoaders;
 using Dynamo.Nodes;
 using Dynamo.Properties;
-using Dynamo.Scheduler.DynamoScheduler;
-using Dynamo.Scheduler.SchedulerThread;
+using Dynamo.Scheduler;
 using Dynamo.Search;
 using Dynamo.Search.SearchElements;
 using Dynamo.Selection;
