+++ src/DynamoCore/Models/HomeWorkspaceModel.cs
@@ -8,9 +8,8 @@ using System.Threading;
 using System.Xml;
 
 using Dynamo.Core;
-using Dynamo.Core.Threading;
 using Dynamo.Engine;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using ProtoCore;
 using ProtoCore.Namespace;
