+++ src/DynamoCore/Scheduler/Disposable.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
-namespace Dynamo.Core.Threading
+namespace Dynamo.Scheduler
 {
     public static class Disposable
     {
