+++ src/DynamoCore/Scheduler/PreviewGraphAsyncTask.cs
@@ -1,17 +1,10 @@
-﻿
-using System;
+﻿using System;
 using System.Collections.Generic;
-
 using Dynamo.Engine;
 using Dynamo.Models;
-using Dynamo.Scheduler.DynamoScheduler;
 using ProtoScript.Runners;
 
-using BuildWarning = ProtoCore.BuildData.WarningEntry;
-using RuntimeWarning = ProtoCore.BuildData.WarningEntry;
-using Dynamo.Interfaces;
-
-namespace Dynamo.Core.Threading
+namespace Dynamo.Scheduler
 {
     class PreviewGraphAsyncTask : AsyncTask
     {
