+++ src/DynamoCore/Scheduler/UpdateGraphAsyncTask.cs
@@ -1,17 +1,12 @@
-﻿using System.Linq;
- using System;
+﻿using System;
 using System.Collections.Generic;
-
+using System.Linq;
 using Dynamo.Engine;
 using Dynamo.Models;
-using Dynamo.Scheduler.DynamoScheduler;
+using ProtoCore.BuildData;
 using ProtoScript.Runners;
 
-using BuildWarning = ProtoCore.BuildData.WarningEntry;
-using RuntimeWarning = ProtoCore.Runtime.WarningEntry;
-using System.Diagnostics;
-
-namespace Dynamo.Core.Threading
+namespace Dynamo.Scheduler
 {
     class UpdateGraphAsyncTask : AsyncTask
     {
@@ -106,8 +101,8 @@ namespace Dynamo.Core.Threading
         {
             if (engineController.IsDisposed)
             {
-                BuildWarnings = new Dictionary<Guid, List<BuildWarning>>();
-                RuntimeWarnings = new Dictionary<Guid, List<RuntimeWarning>>();
+                BuildWarnings = new Dictionary<Guid, List<WarningEntry>>();
+                RuntimeWarnings = new Dictionary<Guid, List<ProtoCore.Runtime.WarningEntry>>();
             }
             else
             {
@@ -186,8 +181,8 @@ namespace Dynamo.Core.Threading
         #region Public Class Properties
 
         internal WorkspaceModel TargetedWorkspace { get; private set; }
-        internal IDictionary<Guid, List<BuildWarning>> BuildWarnings { get; private set; }
-        internal IDictionary<Guid, List<RuntimeWarning>> RuntimeWarnings { get; private set; }
+        internal IDictionary<Guid, List<WarningEntry>> BuildWarnings { get; private set; }
+        internal IDictionary<Guid, List<ProtoCore.Runtime.WarningEntry>> RuntimeWarnings { get; private set; }
 
         #endregion
