+++ src/DynamoCoreWpf/Views/Preview/PreviewControl.xaml.cs
@@ -1,8 +1,7 @@
 ﻿using System.Windows.Threading;
 using Dynamo.Controls;
-using Dynamo.Core.Threading;
 using Dynamo.Interfaces;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using Dynamo.ViewModels;
 using ProtoCore.Mirror;
 using System;
