+++ src/Libraries/CoreNodeModelsWpf/NodeViewCustomizations/Watch.cs
@@ -8,9 +8,9 @@ using System.Windows.Threading;
 using DSCoreNodesUI.Logic;
 using Dynamo.Configuration;
 using Dynamo.Controls;
-using Dynamo.Core.Threading;
 using Dynamo.Interfaces;
 using Dynamo.Models;
+using Dynamo.Scheduler;
 using Dynamo.UI;
 using Dynamo.Nodes;
 using Dynamo.ViewModels;
