+++ src/Libraries/DynamoWatch3D/Watch3D.cs
@@ -11,11 +11,9 @@ using System.Windows.Threading;
 using System.Xml;
 using Autodesk.DesignScript.Interfaces;
 using Dynamo.Controls;
-using Dynamo.Core.Threading;
 using Dynamo.Engine;
-using Dynamo.Interfaces;
 using Dynamo.Models;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using Dynamo.Utilities;
 using Dynamo.Logging;
 using Dynamo.Visualization;
