+++ src/VisualizationTests/HelixWatch3DViewModelTests.cs
@@ -11,7 +11,7 @@ using Dynamo;
 using Dynamo.Controls;
 using Dynamo.Models;
 using Dynamo.Nodes;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using Dynamo.Tests;
 using Dynamo.UI;
 using Dynamo.Utilities;
