+++ test/DynamoCoreTests/AsyncTaskExtensionsTests.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Dynamo.Core.Threading;
-using Dynamo.Scheduler.DynamoScheduler;
+﻿using System.Collections.Generic;
+using Dynamo.Scheduler;
 using NUnit.Framework;
 
 namespace Dynamo.Tests
