+++ test/DynamoCoreTests/DynamoModelTestBase.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Reflection;
 using Dynamo.Interfaces;
 using Dynamo.Models;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using Dynamo.Selection;
 using Dynamo.Tests;
 using Dynamo.Utilities;
