+++ test/DynamoCoreTests/ExtensionTests.cs
@@ -1,15 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using NUnit.Framework;
 using Dynamo.Extensions;
 using System.IO;
 using Dynamo.Models;
 using Moq;
-using Dynamo.Interfaces;
 
 namespace Dynamo.Tests
 {
