+++ test/DynamoCoreTests/SchedulerTests.cs
@@ -1,12 +1,8 @@
 ﻿using System.IO;
 using System.Reflection;
-
-using Dynamo.Core.Threading;
-using Dynamo.Interfaces;
 using Dynamo.Models;
 using Dynamo.Nodes;
-using Dynamo.Scheduler.DynamoScheduler;
-using Dynamo.Scheduler.SchedulerThread;
+using Dynamo.Scheduler;
 using DynamoShapeManager;
 
 using NUnit.Framework;
