+++ test/DynamoCoreWpfTests/DynamoTestUIBase.cs
@@ -4,13 +4,11 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading;
-using Dynamo;
 using Dynamo.Configuration;
 using Dynamo.Controls;
 using Dynamo.Models;
-using Dynamo.Scheduler.DynamoScheduler;
+using Dynamo.Scheduler;
 using Dynamo.ViewModels;
-using Dynamo.Wpf.ViewModels.Watch3D;
 using DynamoShapeManager;
 using NUnit.Framework;
 using TestServices;
